refactor: メソッド分割
