+++ NotionAPI/Sources/Models/Models.cs
@@ -1,5 +1,4 @@
-﻿using System.ComponentModel.DataAnnotations;
-using System.Text.Json.Serialization;
+﻿using System.Text.Json.Serialization;
 
 namespace Notion;
 
@@ -33,16 +32,6 @@ public class User
 }
 
 [Serializable]
-public class UsersAPIResponse
-{
-    [JsonPropertyName("object")]
-    public string Object { get; set; } = string.Empty;
-
-    [JsonPropertyName("results")]
-    public List<User> Results { get; set; } = new List<User>();
-}
-
-[Serializable]
 public class SearchSort
 {
     public const string Ascending = "ascending";
@@ -66,32 +55,6 @@ public class Filter
 
     [JsonPropertyName("value")]
     public string Value { get; set; } = Page;
-}
-
-/// <summary>
-/// Search by title
-/// </summary>
-/// <see href="https://developers.notion.com/reference/post-search"/>
-[Serializable]
-public class SearchAPIRequest
-{
-    public const int Max_Page_Size = 100;
-
-    [JsonPropertyName("query")]
-    public string Query { get; set; } = string.Empty;
-
-    [JsonPropertyName("filter")]
-    public Filter Filter { get; set; } = new();
-
-    [JsonPropertyName("sort")]
-    public SearchSort Sort { get; set; } = new();
-
-    [JsonPropertyName("start_cursor")]
-    public string StartCursor { get; set; } = string.Empty;
-
-    [JsonPropertyName("page_size")]
-    [Range(1, Max_Page_Size)]
-    public int PageSize { get; set; } = Max_Page_Size;
 }
 
 [Serializable]
@@ -262,30 +225,4 @@ public class SearchResult
 
     [JsonPropertyName("public _url")]
     public string PublicURL { get; set; } = string.Empty;
-}
-
-/// <summary>
-/// Search by title
-/// </summary>
-/// <see href="https://developers.notion.com/reference/post-search"/>
-[Serializable]
-public class SearchAPIResponse
-{
-    [JsonPropertyName("object")]
-    public string Object { get; set; } = string.Empty;
-
-    [JsonPropertyName("results")]
-    public List<SearchResult> Results { get; set; } = new List<SearchResult>();
-
-    [JsonPropertyName("next_cursor")]
-    public string NextCursor { get; set; } = string.Empty;
-
-    [JsonPropertyName("has_more")]
-    public bool HasMore { get; set; }
-
-    [JsonPropertyName("type")]
-    public string Type { get; set; } = string.Empty;
-
-    [JsonPropertyName("page_or_database")]
-    public object PageOrDatabase { get; set; }
 }
